requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
pydantic==2.6.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.23
python-dotenv==1.0.0